+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unrealeased [=> 0.4.7]
 ## Added
+- File typings
 - Support for Iterator typing
 - Object's default_dicts now implement objct_class